Incorporated Chapter 4 functions into utils
Incorporated the functions in Chapter_04.py into utils.py
